+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 6.0.0
 
 * Changed ForgotPasswordform to use the same layout as the EmailPasswordLoginForm.
+* Add option for textalignment for email and password fields.
 
 ## 5.2.0